+++ CHANGELOG.md
@@ -2168,3 +2168,4 @@
 - [2026-06-19 23:04] chore(test): update dependencies and cleanup unused imports in test
 - [2026-06-19 23:28] fix(analytics): prevent double tap duplicate action in analytics
 - [2026-06-19 23:29] fix(ui): prevent double tap duplicate action in ui
+- [2026-06-21 09:45] perf(perf): apply RepaintBoundary on animated perf widgets